+++ src/capturer/model.rs
@@ -1,7 +1,5 @@
-use crate::model::Result;
-use windows::Win32::Graphics::Dxgi::{
-  DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTDUPL_POINTER_SHAPE_INFO, DXGI_OUTPUT_DESC,
-};
+use crate::model::{PointerShape, Result};
+use windows::Win32::Graphics::Dxgi::{DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTPUT_DESC};
 
 /// Capturer is stateful, it holds a buffer of the last captured frame.
 pub trait Capturer {
@@ -22,10 +20,7 @@ pub trait Capturer {
   /// The pixel data is stored in the buffer.
   fn capture(&mut self) -> Result<DXGI_OUTDUPL_FRAME_INFO>;
 
-  fn get_pointer(
-    &self,
-    info: &DXGI_OUTDUPL_FRAME_INFO,
-  ) -> Result<(Vec<u8>, u32, DXGI_OUTDUPL_POINTER_SHAPE_INFO)>;
+  fn get_pointer_shape(&self, info: &DXGI_OUTDUPL_FRAME_INFO) -> Result<PointerShape>;
 
   /// Check buffer size before capture.
   fn safe_capture(&mut self) -> Result<DXGI_OUTDUPL_FRAME_INFO>;